+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [4.1.30] - 2022-09-13
 ### Added
 - Add reactions to group posts [#2354](https://github.com/rokwire/illinois-app/issues/2354)
+### Fixed
+- Compound widget style for favorites [#2373](https://github.com/rokwire/illinois-app/issues/2373).
 
 ## [4.1.29] - 2022-09-12
 ### Fixed